backend/requirements.txt
asgiref==3.8.1
Babel==2.14.0
Django==5.0.4
django-filter==24.2
django-money==3.4.1
djangorestframework==3.15.1
pillow==10.3.0
py-moneyed==3.0
pycodestyle==2.11.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
